Stability fixes and memory optimizations, Also fixes up collision hack
This is the first in a wave of a few features and reworks, backported from the C# port. The main features/improvements being:

Clearing out the SurfaceTool arrays when we are done with them, freeing, more usable memory at least on C# (I was able to run a 64x64 grid of 16x255x16 chunks on 5gb of ram, previously took close to 11gb)

Proper threading when dealing with the Scene tree in most areas, except for the removal of chunks due to the performance impact being thread safe has on it, will look for a solution to that problem at a later date.

Finally removal of the hack used in the previous collision setup that will be more useful down the line when dealing with the bug regarding being unable to break blocks without collision (No I did not fix that)